refactor(mcp): make UI builders private
